# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlimefishBackend API",
    "version": "1.0",
    "description": "[Discord](https://discord.gg/Q5CeSMFeBP) · [API key](https://slimefishbackend.dev/settings/api) · [OpenAPI spec](https://github.com/casesandberg/slimefish-backend/blob/main/apps/api/openapi.json) · [Trading bot](https://github.com/casesandberg/slimefishbackend-bot-trading) · [Mirror bot](https://github.com/casesandberg/slimefishbackend-bot-manifold-mirror)\n\n Include `x-api-key: {{key}}` in the header of your requests."
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "query": {
        "type": "string"
      },
...
    },
    "parameters": {
...
      "query": {
        "schema": {
          "$ref": "#/components/schemas/query"
        },
        "required": false,
        "name": "query",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/search": {
      "get": {
        "summary": "Search for users, markets, and lists",
        "parameters": [
          {
            "$ref": "#/components/parameters/query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "role": {
                                "type": "string",
                                "enum": [
                                  "USER",
                                  "ADMIN"
                                ]
                              },
                              "id": {
                                "type": "string",
                                "format": "cuid"
                              },
                              "username": {
                                "type": "string"
                              },
                              "displayName": {
                                "type": "string"
                              },
                              "avatarUrl": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "twitterHandle": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "discordHandle": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "website": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "bio": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "timezone": {
                                "type": "string"
                              },
                              "primaryAccountId": {
                                "type": "string"
                              },
                              "referralCode": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "referredBy": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "createdAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "updatedAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              }
                            },
                            "required": [
                              "role",
                              "id",
                              "username",
                              "displayName",
                              "avatarUrl",
                              "twitterHandle",
                              "discordHandle",
                              "website",
                              "bio",
                              "timezone",
                              "primaryAccountId",
                              "referralCode",
                              "referredBy",
                              "createdAt",
                              "updatedAt"
                            ]
                          }
                        },
                        "markets": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string",
                                "format": "cuid"
                              },
                              "question": {
                                "type": "string",
                                "minLength": 1
                              },
                              "description": {
                                "type": "string"
                              },
                              "slug": {
                                "type": "string",
                                "minLength": 1
                              },
                              "closeDate": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "resolvedAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "canceledAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "canceledById": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "createdBy": {
                                "type": "string"
                              },
                              "tags": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                },
                                "maxItems": 5
                              },
                              "ammAccountId": {
                                "type": "string"
                              },
                              "clearingAccountId": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "updatedAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "commentCount": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "uniqueTradersCount": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "uniquePromotersCount": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "liquidityCount": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "parentListId": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "question",
                              "description",
                              "slug",
                              "closeDate",
                              "resolvedAt",
                              "canceledAt",
                              "canceledById",
                              "createdBy",
                              "tags",
                              "ammAccountId",
                              "clearingAccountId",
                              "createdAt",
                              "updatedAt",
                              "commentCount",
                              "uniqueTradersCount",
                              "uniquePromotersCount",
                              "liquidityCount",
                              "parentListId"
                            ]
                          }
                        },
                        "lists": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "contributionPolicy": {
                                "type": "string",
                                "enum": [
                                  "DISABLED",
                                  "OWNERS_ONLY",
                                  "FRIENDS_ONLY",
                                  "PUBLIC"
                                ]
                              },
                              "id": {
                                "type": "string",
                                "format": "cuid"
                              },
                              "title": {
                                "type": "string",
                                "minLength": 1
                              },
                              "slug": {
                                "type": "string",
                                "minLength": 1
                              },
                              "description": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "ownerId": {
                                "type": "string"
                              },
                              "contributionReview": {
                                "type": [
                                  "boolean",
                                  "null"
                                ]
                              },
                              "tags": {
                                "type": "array",
                                "items": {
                                  "type": "string"
                                },
                                "maxItems": 5
                              },
                              "createdAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "updatedAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              }
                            },
                            "required": [
                              "contributionPolicy",
                              "id",
                              "title",
                              "slug",
                              "description",
                              "ownerId",
                              "contributionReview",
                              "tags",
                              "createdAt",
                              "updatedAt"
                            ]
                          }
                        }
                      },
                      "required": [
                        "users",
                        "markets",
                        "lists"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}